requirements.txt
python-dotenv
requests
pyTelegramBotAPI
aiohttp
Jinja2
tenacity

